feature/dev: add Private module
